# specs/api.yml
openapi: 3.0.0
info:
  title: OnePage 電商平台 API
  version: 1.0.0
  description: |
    OnePage 一頁式電商平台 API 契約。
    後端：Spring Boot 3.2 + JPA + PostgreSQL。
    金流：綠界全方位金流 API v5。
    發票：綠界電子發票 API。
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /websites:
    get:
      tags: [Website]
      summary: 查詢網站清單
      description: 管理員返回所有網站；一般使用者僅返回自己建立的網站（REQ-034 RBAC 資料隔離）
      operationId: listWebsites
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 網站清單（依 RBAC 角色過濾）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebsiteDTO'

    post:
      tags: [Website]
      summary: 建立網站
      operationId: createWebsite
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateWebsiteRequest'
      responses:
        '201':
          description: 網站建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteDTO'
        '400':
          description: 驗證失敗
  /websites/{websiteId}:
    get:
      tags: [Website]
      summary: 查詢單一網站
      operationId: getWebsite
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 網站詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteDTO'
        '404':
          description: 網站不存在

    put:
      tags: [Website]
      summary: 編輯網站
      operationId: updateWebsite
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateWebsiteRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteDTO'
        '400':
          description: 驗證失敗
  /websites/{websiteId}/publish:
    post:
      tags: [Website]
      summary: 啟用網站（草稿→已上線）
      operationId: publishWebsite
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 啟用成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteDTO'
        '400':
          description: 狀態轉換不合法
  /websites/{websiteId}/unpublish:
    post:
      tags: [Website]
      summary: 停用網站（已上線→已下線）
      operationId: unpublishWebsite
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 停用成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteDTO'
        '400':
          description: 狀態轉換不合法
  /websites/{websiteId}/republish:
    post:
      tags: [Website]
      summary: 重新上線網站（已下線→已上線）（REQ-028）
      operationId: republishWebsite
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 重新上線成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteDTO'
        '400':
          description: 狀態轉換不合法（僅已下線網站可重新上線）
        '403':
          description: 無權限（非 owner 且非管理員）
        '404':
          description: 網站不存在
  /websites/{websiteId}/products:
    get:
      tags: [Website]
      summary: 查詢網站已上架商品
      operationId: listWebsiteProducts
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 網站商品清單
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebsiteProductDTO'

    put:
      tags: [Website]
      summary: 批次更新網站上架商品（覆蓋式）
      operationId: updateWebsiteProducts
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/WebsiteProductInput'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebsiteProductDTO'
        '400':
          description: 商品不存在或非已上架狀態

  # ═══ Storefront（前台） ═══
  /storefront/websites/{websiteId}:
    get:
      tags: [Storefront]
      summary: 取得網站前台頁面資料（集合頁）
      operationId: getStorefrontWebsite
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 網站前台資料（含商品列表）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorefrontWebsiteDTO'
        '404':
          description: 網站不存在或未上線
  /storefront/websites/{websiteId}/products/{productSlug}:
    get:
      tags: [Storefront]
      summary: 取得商品詳情頁資料
      operationId: getStorefrontProduct
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: productSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 商品詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorefrontProductDTO'
        '404':
          description: 商品不存在或未上架至此網站
  /storefront/products/{productId}/stock-check:
    get:
      tags: [Storefront]
      summary: 檢查商品庫存是否足夠（加入購物車前）
      operationId: checkProductStock
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: quantity
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: 庫存檢查結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
        '404':
          description: 商品不存在

  # ═══ Shipping ═══
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BundleItemDTO:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        productName:
          type: string
        price:
          type: number
          format: double
    ProductImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        imageUrl:
          type: string
        sortOrder:
          type: integer
...
    WebsiteDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        title:
          type: string
          description: 網站標題（前台顯示用）（REQ-029）
        subtitle:
          type: string
          nullable: true
          description: 網站副標題（選填）（REQ-029）
        browserTitle:
          type: string
          nullable: true
          description: 瀏覽器分頁標題 <title> tag（選填）（REQ-029）
        footerTitle:
          type: string
          nullable: true
          description: Footer 標題（店名）（選填）（REQ-029）
        footerSubtitle:
          type: string
          nullable: true
          description: Footer 副標題（聯絡資訊）（選填）（REQ-029）
        subscriptionPlan:
          type: string
        publishStartAt:
          type: string
          format: date-time
          nullable: true
        publishEndAt:
          type: string
          format: date-time
          nullable: true
        bannerImageUrl:
          type: string
          nullable: true
        promoImageUrl:
          type: string
          nullable: true
        freeShippingThreshold:
          type: number
          format: double
        status:
          type: string
          enum: [DRAFT, PUBLISHED, OFFLINE]
        storefrontUrl:
          type: string
          description: 前台 URL /project/{websiteId}
        ownerUserId:
          type: integer
          format: int64
          description: 建立者 user ID（RBAC 資料隔離用）（REQ-034）
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateWebsiteRequest:
      type: object
      required: [name, title, bannerImage]
      properties:
        name:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
          description: 網站標題（必填）（REQ-029）
        subtitle:
          type: string
          nullable: true
          description: 網站副標題（選填）（REQ-029）
        browserTitle:
          type: string
          nullable: true
          description: 瀏覽器分頁標題（選填）（REQ-029）
        footerTitle:
          type: string
          nullable: true
          description: Footer 標題（選填）（REQ-029）
        footerSubtitle:
          type: string
          nullable: true
          description: Footer 副標題（選填）（REQ-029）
        subscriptionPlan:
          type: string
        publishStartAt:
          type: string
          format: date-time
        publishEndAt:
          type: string
          format: date-time
        bannerImage:
          type: string
          format: binary
          description: Banner 主視覺圖片（必填）（REQ-029）
        promoImage:
          type: string
          format: binary
          nullable: true
          description: 宣傳文案圖片（選填）（REQ-029）
        freeShippingThreshold:
          type: number
          format: double
          default: 1500.00
    UpdateWebsiteRequest:
      type: object
      required: [name, title]
      properties:
        name:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
          description: 網站標題（必填）（REQ-029）
        subtitle:
          type: string
          nullable: true
          description: 網站副標題（選填）（REQ-029）
        browserTitle:
          type: string
          nullable: true
          description: 瀏覽器分頁標題（選填）（REQ-029）
        footerTitle:
          type: string
          nullable: true
          description: Footer 標題（選填）（REQ-029）
        footerSubtitle:
          type: string
          nullable: true
          description: Footer 副標題（選填）（REQ-029）
        subscriptionPlan:
          type: string
        publishStartAt:
          type: string
          format: date-time
        publishEndAt:
          type: string
          format: date-time
        bannerImage:
          type: string
          format: binary
          nullable: true
          description: Banner 主視覺圖片（REQ-029）
        promoImage:
          type: string
          format: binary
          nullable: true
          description: 宣傳文案圖片（選填）（REQ-029）
        freeShippingThreshold:
          type: number
          format: double
    WebsiteProductDTO:
      type: object
      properties:
        websiteId:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        productName:
          type: string
        productSlug:
          type: string
        publishAt:
          type: string
          format: date-time
    WebsiteProductInput:
      type: object
      required: [productId, publishAt]
      properties:
        productId:
          type: integer
          format: int64
        publishAt:
          type: string
          format: date-time

    # ── Storefront ──
    StorefrontWebsiteDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        title:
          type: string
          description: 網站標題（REQ-029）
        subtitle:
          type: string
          nullable: true
          description: 網站副標題（REQ-029）
        browserTitle:
          type: string
          nullable: true
          description: 瀏覽器分頁標題（REQ-029）
        footerTitle:
          type: string
          nullable: true
          description: Footer 標題（REQ-029）
        footerSubtitle:
          type: string
          nullable: true
          description: Footer 副標題（REQ-029）
        bannerImageUrl:
          type: string
          nullable: true
        promoImageUrl:
          type: string
          nullable: true
        freeShippingThreshold:
          type: number
          format: double
        products:
          type: array
          items:
            $ref: '#/components/schemas/StorefrontProductCardDTO'
          description: 已到達上架時間且已關聯至本網站的商品列表（REQ-030）
    StorefrontProductCardDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        price:
          type: number
          format: double
        imageUrl:
          type: string
          nullable: true
        isPreorder:
          type: boolean
        preorderDiscountPercent:
          type: number
          format: double
          nullable: true
    StorefrontProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        priceUnit:
          type: string
          enum: [KG, CATTY]
        packaging:
          type: string
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImageDTO'
        isPreorder:
          type: boolean
        preorderStartDate:
          type: string
          format: date
          nullable: true
        preorderEndDate:
          type: string
          format: date
          nullable: true
        preorderDiscountPercent:
          type: number
          format: double
          nullable: true
        isBundle:
          type: boolean
        bundleDiscountPercent:
          type: number
          format: double
          nullable: true
        bundleItems:
          type: array
          items:
            $ref: '#/components/schemas/BundleItemDTO'
        websiteId:
          type: integer
          format: int64

    # ── Shipping ──
...